LayoutTests/http/tests/inspector/network/local-resource-override-script-tag-expected.txt

ALERT: DEFAULT override.js TEXT
ALERT: OVERRIDDEN override.js TEXT
Test a LocalResourceOverride overriding a <script> load.


== Running test suite: LocalResourceOverride.Script
-- Running test case: LocalResourceOverride.Script.Tag
